CoordFixR: Coordinate Converter Web Application

This package calculates or converts coordinates of (nearly) all possible notations into decimal degrees. It does not matter whether degree, minute or second signs are included (e.g. '°'), whether spaces, underscores or slashes occur, whether designations such as north, south, west and east or minus signs (“-”) are included. Coordinates that contain text components such as “degrees”, “minutes” or “seconds” are also converted correctly without any problems.

Author

Maintainer: Hendrik Pehlke hendrikpehlke@awi.de (ORCID)